removed obsolete methods, extracted classes to their own files
